Keep Parcel size stable.
The size of 32-bit on arm 'Parcel' was accidentally changed. Fix this by
making stability 16-bits (there is not enough space in Parcel right now
to keep the 32-bit mask without combining some of the bools). This broke
a device in internal master with some 32-bit prebuilts.

Bug: 137348710
Test: TH
Test: boots on hawk

+++ include/binder/Stability.h
@@ -45,7 +45,7 @@ private:
     // up the efficiency level of a binder object. So, we expose the underlying type.
     friend ::android::Parcel;
 
-    enum Level : int32_t {
+    enum Level : int16_t {
         UNDECLARED = 0,
 
         VENDOR = 0b000011,
